requirements.txt
python-binance
plotly
pandas
numpy
pandas_ta